fix: fix a heap of issues with SW exceptions
- The logic was convoluted. If there is a pending HW interrupt, a cause value should be forced, otherwise the stored one should be used

+++ sirc-vm/peripheral-cpu/src/coprocessors/exception_unit/definitions.rs
@@ -15,19 +15,19 @@ pub mod vectors {
     /// and another chip detects this and raises an error.
     /// It could also be used to implement virtual memory as the instrtuction is aborted
     /// and pl is not incremented for a bus fault.
-    pub const BUS_FAULT: u16 = 0x00;
+    pub const BUS_FAULT: u8 = 0x00;
     /// Raised when fetching instructions if pl points to an odd address
     /// This is to simplify fetching as the second word of an instruction is always at pl | 0x1
     /// and ensures that we never have to worry about instructions overflowing segments
     /// (e.g. if the first word is at 0xFFFF and the second word at 0x0000)
     /// which is a weird edge case that might complicate fetching.
-    pub const ALIGNMENT_FAULT: u16 = 0x01;
+    pub const ALIGNMENT_FAULT: u8 = 0x01;
     /// Raised with some instructions if the computed address would go outside the current
     /// segment. E.g. if you are accessing data in the stack segment, and then compute
     /// an address that overflows, it is probably a stack overflow.
     /// There might be situations where you want address calculations to wrap around
     /// so it is only raised if the `TrapOnAddressOverflow` SR bit is set.
-    pub const SEGMENT_OVERFLOW_FAULT: u16 = 0x02;
+    pub const SEGMENT_OVERFLOW_FAULT: u8 = 0x02;
     /// Raised when a co-processor call is done for a non-existant co-processor
     /// or if the co-processor opcode is invalid.
     /// Can be used for forward compatibility.
@@ -38,13 +38,13 @@ pub mod vectors {
     /// This is just to keep things simple. There is a risk of people using undocumented instructions
     /// and those programs breaking in future iterations of the CPU but it is expected that the core
     /// of the CPU will remain stable and future ISA improvements will be done via co-processors.
-    pub const INVALID_OPCODE_FAULT: u16 = 0x03;
+    pub const INVALID_OPCODE_FAULT: u8 = 0x03;
 
     /// Raised when not in system mode and a privileged operation is performed:
     /// 1. Writing to the high word of any address registers
     /// 2. Writing to the high byte of the SR register
     /// 3. Triggering exception below 0x80
-    pub const PRIVILEGE_VIOLATION_FAULT: u16 = 0x04;
+    pub const PRIVILEGE_VIOLATION_FAULT: u8 = 0x04;
 
     // 0x05-0x07 Reserved
 
@@ -52,7 +52,7 @@ pub mod vectors {
 
     /// Raised after every instruction when the `TraceMode` SR bit is set
     /// Used for debugging
-    pub const INSTRUCTION_TRACE_FAULT: u16 = 0x08;
+    pub const INSTRUCTION_TRACE_FAULT: u8 = 0x08;
 
     /// Raised when a level five hardware exception is raised
     /// when one is already being handled
@@ -62,7 +62,7 @@ pub mod vectors {
     /// masked, but it could indicate a hardware misconfiguration,
     /// so it is handy so that hardware bugs for things that should
     /// not be interrupted are picked up.
-    pub const LEVEL_FIVE_HARDWARE_EXCEPTION_CONFLICT: u16 = 0x09;
+    pub const LEVEL_FIVE_HARDWARE_EXCEPTION_CONFLICT: u8 = 0x09;
 
     //  0x09-0x0F Reserved
 
@@ -71,17 +71,17 @@ pub mod vectors {
     // Special level - When level five hardware exception is masked and
     // another one is triggered, it isn't ignored, it triggers a LEVEL_FIVE_HARDWARE_EXCEPTION_CONFLICT
     // (see above)
-    pub const LEVEL_FIVE_HARDWARE_EXCEPTION: u16 = 0x10;
-    pub const LEVEL_FOUR_HARDWARE_EXCEPTION: u16 = 0x20;
-    pub const LEVEL_THREE_HARDWARE_EXCEPTION: u16 = 0x30;
-    pub const LEVEL_TWO_HARDWARE_EXCEPTION: u16 = 0x40;
-    pub const LEVEL_ONE_HARDWARE_EXCEPTION: u16 = 0x50;
+    pub const LEVEL_FIVE_HARDWARE_EXCEPTION: u8 = 0x10;
+    pub const LEVEL_FOUR_HARDWARE_EXCEPTION: u8 = 0x20;
+    pub const LEVEL_THREE_HARDWARE_EXCEPTION: u8 = 0x30;
+    pub const LEVEL_TWO_HARDWARE_EXCEPTION: u8 = 0x40;
+    pub const LEVEL_ONE_HARDWARE_EXCEPTION: u8 = 0x50;
 
     /// User Exceptions
     // 159 user exception vectors triggered by the EXCP instruction (e.g. a TRAP on the 68k)
 
-    pub const USER_EXCEPTION_VECTOR_START: u16 = 0x60;
-    pub const USER_EXCEPTION_VECTOR_END: u16 = 0xFF;
+    pub const USER_EXCEPTION_VECTOR_START: u8 = 0x60;
+    pub const USER_EXCEPTION_VECTOR_END: u8 = 0xFF;
 }
 
 // Instruction priorities
@@ -114,8 +114,8 @@ pub enum ExceptionPriorities {
     Fault = 0x7,
 }
 
-#[derive(Debug, PartialEq, Eq, FromPrimitive)]
-
+#[repr(u8)]
+#[derive(Debug, PartialEq, Eq, FromPrimitive, ToPrimitive)]
 pub enum ExceptionUnitOpCodes {
     SoftwareException = 0x1,
